xmodem/requirements.txt
xmodem==0.4.7
pyserial==3.5
